lab5/gender-age-detection/requirements.txt
streamlit
numpy
opencv-python
keras
matplotlib
pillow
tensorflow